# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Infrastructure Copilot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations:
    get:
      operationId: listOrganizations
      tags: [organizations]
      summary: List organizations for current user
      responses:
        "200":
          description: List of organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
    post:
      operationId: createOrganization
      tags: [organizations]
      summary: Create organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /organizations/{orgId}:
    get:
      operationId: getOrganization
      tags: [organizations]
      summary: Get organization
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    patch:
      operationId: updateOrganization
      tags: [organizations]
      summary: Update organization
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    delete:
      operationId: deleteOrganization
      tags: [organizations]
      summary: Delete organization
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /organizations/{orgId}/stats:
    get:
      operationId: getOrganizationStats
      tags: [organizations]
      summary: Get organization overview stats
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgStats"
  /organizations/{orgId}/members:
    get:
      operationId: listOrgMembers
      tags: [organizations]
      summary: List organization members
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrgMember"
    post:
      operationId: addOrgMember
      tags: [organizations]
      summary: Add member to organization
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrgMemberInput"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgMember"
  /organizations/{orgId}/members/{userId}:
    patch:
      operationId: updateOrgMember
      tags: [organizations]
      summary: Update member role
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrgMemberUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgMember"
    delete:
      operationId: removeOrgMember
      tags: [organizations]
      summary: Remove member from organization
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ─── SERVERS ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    OrganizationInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        logoUrl:
          type: string
    OrganizationUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        logoUrl:
          type: string
    Organization:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        memberCount:
          type: integer
        serverCount:
          type: integer
    OrgStats:
      type: object
      required: [totalServers, onlineServers, activeDeployments, failedDeployments, totalDeployments, pendingApprovals, alertCount]
      properties:
        totalServers:
          type: integer
        onlineServers:
          type: integer
        offlineServers:
          type: integer
        activeDeployments:
          type: integer
        failedDeployments:
          type: integer
        totalDeployments:
          type: integer
        pendingApprovals:
          type: integer
        alertCount:
          type: integer
        successRate:
          type: number
    OrgMemberInput:
      type: object
      required: [email, role]
      properties:
        email:
          type: string

        role:
          type: string
          enum: [engineer, reviewer, admin]
    OrgMemberUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [engineer, reviewer, admin]
    OrgMember:
      type: object
      required: [userId, name, email, role, joinedAt]
      properties:
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [engineer, reviewer, admin]
        joinedAt:
          type: string
          format: date-time

    # ─── SERVERS ───────────────────────────────────────────────────────────────
...